+++ xbmc/utils/AlarmClock.h
@@ -48,10 +48,12 @@ public:
     std::map<std::string,SAlarmClockEvent>::iterator iter;
     if ((iter=m_event.find(strName)) != m_event.end())
     {
-      return iter->second.m_fSecs-(iter->second.watch.IsRunning() ? iter->second.watch.GetElapsedSeconds() : 0.f);
+      return iter->second.m_fSecs - static_cast<double>(iter->second.watch.IsRunning()
+                                                            ? iter->second.watch.GetElapsedSeconds()
+                                                            : 0.f);
     }
 
-    return 0.f;
+    return 0.0;
   }
 
   void Stop(const std::string& strName, bool bSilent = false);
